+++ src/OpenTelemetry/Metrics/MetricReaderExt.cs
@@ -28,7 +28,7 @@ namespace OpenTelemetry.Metrics
     public abstract partial class MetricReader
     {
         private readonly HashSet<string> metricStreamNames = new(StringComparer.OrdinalIgnoreCase);
-        private readonly ConcurrentDictionary<InstrumentIdentity, Metric> instrumentIdentityToMetric = new();
+        private readonly ConcurrentDictionary<MetricStreamIdentity, Metric> instrumentIdentityToMetric = new();
         private readonly object instrumentCreationLock = new();
         private int maxMetricStreams;
         private int maxMetricPointsPerMetricStream;
@@ -42,10 +42,10 @@ namespace OpenTelemetry.Metrics
             var meterVersion = instrument.Meter.Version;
             var metricName = instrument.Name;
             var metricStreamName = $"{meterName}.{meterVersion}.{metricName}";
-            var instrumentIdentity = new InstrumentIdentity(instrument.Meter, metricName, instrument.Unit, instrument.Description, instrument.GetType());
+            var metricStreamIdentity = new MetricStreamIdentity(instrument.Meter, metricName, instrument.Unit, instrument.Description, instrument.GetType(), null, null);
             lock (this.instrumentCreationLock)
             {
-                if (this.instrumentIdentityToMetric.TryGetValue(instrumentIdentity, out var existingMetric))
+                if (this.instrumentIdentityToMetric.TryGetValue(metricStreamIdentity, out var existingMetric))
                 {
                     return existingMetric;
                 }
@@ -70,7 +70,7 @@ namespace OpenTelemetry.Metrics
                     Metric metric = null;
                     try
                     {
-                        metric = new Metric(instrumentIdentity, this.Temporality, this.maxMetricPointsPerMetricStream);
+                        metric = new Metric(metricStreamIdentity, this.Temporality, this.maxMetricPointsPerMetricStream);
                     }
                     catch (NotSupportedException nse)
                     {
@@ -82,7 +82,7 @@ namespace OpenTelemetry.Metrics
                         return null;
                     }
 
-                    this.instrumentIdentityToMetric[instrumentIdentity] = metric;
+                    this.instrumentIdentityToMetric[metricStreamIdentity] = metric;
                     this.metrics[index] = metric;
                     this.metricStreamNames.Add(metricStreamName);
                     return metric;
@@ -119,7 +119,10 @@ namespace OpenTelemetry.Metrics
                     var metricName = metricStreamConfig?.Name ?? instrument.Name;
                     var metricStreamName = $"{meterName}.{meterVersion}.{metricName}";
                     var metricDescription = metricStreamConfig?.Description ?? instrument.Description;
-                    var instrumentIdentity = new InstrumentIdentity(instrument.Meter, metricName, instrument.Unit, metricDescription, instrument.GetType());
+                    var tagKeysInteresting = metricStreamConfig?.CopiedTagKeys;
+                    var histogramBucketBounds = (metricStreamConfig is ExplicitBucketHistogramConfiguration histogramConfig
+                        && histogramConfig.CopiedBoundaries != null) ? histogramConfig.CopiedBoundaries : null;
+                    var metricStreamIdentity = new MetricStreamIdentity(instrument.Meter, metricName, instrument.Unit, metricDescription, instrument.GetType(), tagKeysInteresting, histogramBucketBounds);
 
                     if (!MeterProviderBuilderSdk.IsValidInstrumentName(metricName))
                     {
@@ -132,9 +135,15 @@ namespace OpenTelemetry.Metrics
                         continue;
                     }
 
-                    if (this.instrumentIdentityToMetric.TryGetValue(instrumentIdentity, out var existingMetric))
+                    if (this.instrumentIdentityToMetric.TryGetValue(metricStreamIdentity, out var existingMetric))
                     {
-                        metrics.Add(existingMetric);
+                        // The list of metrics may already contain a matching metric with the same
+                        // identity when a single instrument is selected by multiple views.
+                        if (!metrics.Contains(existingMetric))
+                        {
+                            metrics.Add(existingMetric);
+                        }
+
                         continue;
                     }
 
@@ -161,12 +170,9 @@ namespace OpenTelemetry.Metrics
                     else
                     {
                         Metric metric;
-                        string[] tagKeysInteresting = metricStreamConfig?.CopiedTagKeys;
-                        double[] histogramBucketBounds = (metricStreamConfig is ExplicitBucketHistogramConfiguration histogramConfig
-                            && histogramConfig.CopiedBoundaries != null) ? histogramConfig.CopiedBoundaries : null;
-                        metric = new Metric(instrumentIdentity, this.Temporality, this.maxMetricPointsPerMetricStream, histogramBucketBounds, tagKeysInteresting);
+                        metric = new Metric(metricStreamIdentity, this.Temporality, this.maxMetricPointsPerMetricStream, histogramBucketBounds, tagKeysInteresting);
 
-                        this.instrumentIdentityToMetric[instrumentIdentity] = metric;
+                        this.instrumentIdentityToMetric[metricStreamIdentity] = metric;
                         this.metrics[index] = metric;
                         metrics.Add(metric);
                         this.metricStreamNames.Add(metricStreamName);