+++ src/OpenTelemetry/CHANGELOG.md
@@ -17,6 +17,10 @@ Released 2022-Mar-30
   period.
   ([#2982](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2982))
 
+* Fix bug where multiple views selecting a single instrument can result in
+  duplicate updates to a single metric point.
+  ([#3006](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3006))
+
 * Added the `PeriodicExportingMetricReaderOptions.ExportTimeoutMilliseconds`
   option.
   ([#3038](https://github.com/open-telemetry/opentelemetry-dotnet/pull/3038))